ciao-cert/ssntp: Migrate cert creation functions to new package
The new ssntp/certs package now contains the functions for creating
client and server certificates. This change only migrates the functions
to a new package and renames them to export them. The function
implementations are not changed.
